// packages/core/src/schema.generated.ts
export const schemas = {
  "0.1": {
// ...
    "properties": {
      "cvd_policy": {
        "const": "0.1"
      },
      "canonical": {
        "type": "string",
        "format": "uri",
        "pattern": "^https://"
      },
// ...
    },
// ...
  },
// ...
} as const;

export const LATEST_VERSION = "0.2";